requirements.txt
streamlit
pandas
st-aggrid
pillow
requests